feat: added skip column to bot table

+++ api/models.py
@@ -126,7 +126,7 @@ class BotModel(DateAbtract):
     pip_margin = models.DecimalField(decimal_places=2, max_digits=4)
     side = models.CharField(max_length=5, choices=(
         ("buy", "buy"), ("sell", "sell"), ("all", "all")), default="all")
-    skip = models.IntegerField(default=0)
+    skip = models.DecimalField(decimal_places=2, max_digits=6, default=0)
 
     def __str__(self):
         return f"{self.owner.username} {self.symbol}"
